Disable gateway if cart is less than 50 cents
Stripe can only process an order that is at least 50 cents, people are
mentioning that the order fails if the cart total is small. So instead
of giving customers an error message for something that they can’t
control, I think disabling the gateway is the better course of action.